requirements.txt
alabaster==1.0.0
babel==2.17.0
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.0
distlib==0.3.9
docutils==0.21.2
et_xmlfile==2.0.0
filelock==3.18.0
flake8==7.2.0
Flask==3.1.1
Flask-DebugToolbar==0.16.0
gunicorn==23.0.0
identify==2.6.12
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.16.0
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
pycodestyle==2.13.0
pyflakes==3.3.2
Pygments==2.19.1
pytest==8.4.0
pytest-cov==6.2.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
roman-numerals-py==3.1.0
six==1.17.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
watchdog==6.0.0
Werkzeug==3.1.3
gunicorn==21.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
